Move utils to models file
